fs.walk() in the _libraryFiles method in validator loads `graceful-fs` before `fs` and ends up falling back to the real filesystem on linux.  We move the validation tests to use the real filesystem as a workaround.
