+++ test/librepo_fs.spec.js
@@ -17,8 +17,11 @@
  ******************************************************************************
  */
 
-import {expect} from './test-setup';
+import {expect, sinon} from './test-setup';
 import {FileSystemNamingStrategy, FileSystemLibraryRepository} from '../src/librepo_fs';
+import {LibraryContributor} from '../src/libcontribute';
+import {makeCompleteV2Lib} from './librepo_fs_mock.spec';
+import {makeTestLib} from './librepo_fs_mock.spec';
 const fs = require('fs');
 const path = require('path');
 
@@ -116,6 +119,86 @@ describe('File System', () => {
 
 	it('can migrate a v1 library without examples to v2 format', () => {
 		return assertMigrate('library-v1-noexamples', 'library-v2-noexamples');
+	});
+
+
+	describe('LibraryContributor', () => {
+
+		it('fails if the library does not validate', () => {
+			const name = 'fred';
+			const client = undefined;
+			const tmpobj = tmp.dirSync();
+			const dir = tmpobj.name;
+			const repo = new FileSystemLibraryRepository(dir);
+			const lib = makeTestLib(name, '1.2.3');
+			const sut = new LibraryContributor({repo, client});
+
+			sut._contribute = sinon.stub();
+
+			const result = repo.add(lib, 2)
+				.then(() => sut.contribute(() => {}, name));
+			return expect(result).to.eventually.be.rejected;
+		});
+
+		it('can contribute a library as a tarball', () => {
+			const name = 'fred';
+			const client = undefined;
+
+			const tmpobj = tmp.dirSync();
+			const dir = tmpobj.name;
+			const repo = new FileSystemLibraryRepository(dir);
+			const lib = makeCompleteV2Lib(name, '1.2.3');
+
+			const sut = new LibraryContributor({repo, client});
+			const callback = sinon.stub();
+			sut._contribute = sinon.stub();
+
+			const result = repo.add(lib, 2)
+				.then(() => sut.contribute(callback, name))
+				.then(() => {
+					expect(sut._contribute).to.have.been.calledOnce;
+					expect(sut._contribute).to.have.been.calledWith(name);
+					const pipe = sut._contribute.firstCall.args[1];
+
+					const zlib = require('zlib');
+					const tar = require('tar-stream');
+					const gunzip = zlib.createGunzip();
+					const extract = tar.extract();
+					const names = [];
+					extract.on('entry', (header, stream, callback) => {
+						names.push(header.name);
+						stream.on('end', () => callback());
+						stream.resume();
+					});
+					const promise = new Promise((fulfill, reject) => {
+						extract.on('finish', fulfill);
+						extract.on('error', reject);
+						pipe.pipe(gunzip).pipe(extract);
+					});
+					return promise.then(() => {
+						expect(names).to.include('README.md');
+						expect(names).include('LICENSE.');
+						expect(names).include('library.properties');
+						expect(names).include('src/fred.cpp');
+						expect(names).include('src/fred.h');
+
+						expect(callback).to.have.been.calledWith('validatingLibrary');
+						expect(callback).to.have.been.calledWith('contributingLibrary');
+						expect(callback).to.have.been.calledWith('contributeComplete');
+					});
+				});
+			return result;
+		});
+
+		it('can attempt to publish from a repo for real', () => {
+			const repo = new FileSystemLibraryRepository('mydir');
+			return expect(repo.contribute('abcd', {}, false, () => {})).to.eventually.be.rejected;
+		});
+
+		it('can attempt to publish from a repo as a dry run', () => {
+			const repo = new FileSystemLibraryRepository('mydir');
+			return expect(repo.contribute('abcd', {}, true, () => arguments[1])).to.eventually.be.rejected;
+		});
 	});
 
 });
